# pixles-sdk/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pixles API",
    "description": "Pixles API Documentation",
    "license": {
      "name": "GNU Affero General Public License v3.0 or later"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server URL"
    }
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "tags": [
          "auth",
          "Pixles Authentication API"
        ],
        "summary": "Get user profile",
        "operationId": "get_user_profile",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "auth",
          "Pixles Authentication API"
        ],
        "summary": "Update user profile",
        "operationId": "update_user_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid password"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthError": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "InvalidCredentials"
            ]
          },
          {
            "type": "string",
            "enum": [
              "UserAlreadyExists"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "$ref": "#/components/schemas/BadRegisterUserRequestError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "InvalidToken"
            ],
            "properties": {
              "InvalidToken": {
                "$ref": "#/components/schemas/ClaimValidationError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "InternalServerError"
            ],
            "properties": {
              "InternalServerError": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Authentication error"
      },
      "BadRegisterUserRequestError": {
        "type": "string",
        "enum": [
          "Invalid email",
          "Invalid username",
          "Invalid password"
        ]
      },
      "ClaimValidationError": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "TokenMissing"
            ]
          },
          {
            "type": "string",
            "enum": [
              "TokenExpired"
            ]
          },
          {
            "type": "object",
            "required": [
              "TokenInvalid"
            ],
            "properties": {
              "TokenInvalid": {
                "type": "string"
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "UnexpectedHeaderFormat"
            ]
          },
          {
            "type": "string",
            "enum": [
              "InvalidScopes"
            ]
          }
        ]
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "new_password": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "email",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "InternalServerError": {
        "description": "Internal server error"
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}